configured the SSR functionality
