+++ package.json
@@ -33,12 +33,16 @@
     "css-loader": "^3.5.3",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "history": "^4.10.1",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
+    "ignore-styles": "^5.0.1",
     "mini-css-extract-plugin": "^0.7.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
+    "react-router": "^5.2.0",
+    "react-router-config": "^5.1.1",
     "react-router-dom": "^5.2.0",
     "react-stripe-checkout": "^2.6.3",
     "react-toastify": "^6.0.5",
