+++ src/SSR-server/index.js
@@ -1,3 +1,5 @@
+require('ignore-styles');
+
 require('@babel/register')({
     presets: ['@babel/preset-env', '@babel/preset-react'],
 });
