+++ src/SSR-server/server.js
@@ -1,6 +1,18 @@
 const express = require('express');
 const webpack = require('webpack');
 
+import React from 'react';
+import { renderToString } from 'react-dom/server';
+import { Provider } from 'react-redux';
+import { createStore } from 'redux';
+import { renderRoutes } from 'react-router-config';
+import { StaticRouter } from 'react-router-dom';
+import reducer from '../frontend/reducers';
+import initialState from '../frontend/initialState'; 
+
+import serverRoutes from '../frontend/routes/ServerRoutes';
+
+
 //const passport = require('passport');
 //const boom = require('@hapi/boom');
 //const cookieParser = require('cookie-parser');
@@ -9,6 +21,7 @@ const webpack = require('webpack');
 const { config } = require("../config/index");
 const { ENV, PORT } = process.env;
 const app = express(); 
+
 
 if(ENV === 'development'){
     console.log('development config');
@@ -23,27 +36,44 @@ if(ENV === 'development'){
 
 }
 
+const setResponse = (html) => {
+  return (`
+    <!DOCTYPE html>
+    <html lang="es">
+      <head>
+        <meta charset="UTF-8">
+        <meta http-equiv="X-UA-Compatible" content="ie=edge">
+        <link rel="stylesheet" href="assets/app.css" type="text/css">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <meta http-equiv="X-UA-Compatible" content="ie=edge">
+        <title>Kod3rs Store</title>
+        
+      </head>
+      <body>
+        <div id="app">${html}</div>
+        <script src="assets/app.js" type="text/javascript"> </script>
+        <script src="https://kit.fontawesome.com/473d269aa9.js"></script>
+      </body>
+    </html>
+  `)
+}
 
-app.get('*', (req, res) => {
-  res.send(`<!DOCTYPE html>
-            <html lang="es">
-              <head>
-                <meta charset="UTF-8">
-                <meta http-equiv="X-UA-Compatible" content="ie=edge">
-                <link rel="stylesheet" href="assets/app.css" type="text/css">
-                <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                <meta http-equiv="X-UA-Compatible" content="ie=edge">
-                <title>Kod3rs Store</title>
-                
-              </head>
-              <body>
-                <div id="app"></div>
-                <script src="assets/app.js" type="text/javascript"> </script>
-                <script src="https://kit.fontawesome.com/473d269aa9.js"></script>
-              </body>
-            </html>
-  `);
-});
+const renderApp = (req, res) => {
+  const store = createStore(reducer, initialState);
+  const html = renderToString(
+    <Provider store={store}>
+      <StaticRouter location={req.url} context={{}}>
+        {renderRoutes(serverRoutes)}
+      </StaticRouter>
+    </Provider>
+  );
+  res.send(setResponse(html));
+}
+
+
+
+
+app.get('*', renderApp);
 
 app.listen(PORT, (err) => {
   if(err) console.error(err);
